Use module files.
